Use ptrdiff_t, not int for signed version of sizeof().

+++ joe/types.h
@@ -3,7 +3,7 @@
 #define _FILE_OFFSET_BITS 64
 #define TO_DIFF_OK(a) ((ptrdiff_t)(a)) /* Means it's OK that we are converting off_t to ptrdiff_t in this case */
 #define TO_CHAR_OK(a) ((char)(a)) /* Means it's OK that we are converting int to char */
-#define SIZEOF(a) ((int)sizeof(a)) /* Signed version of sizeof() */
+#define SIZEOF(a) ((ptrdiff_t)sizeof(a)) /* Signed version of sizeof() */
 
 #define WIND_BW(x, y) do { \
   if (!((y)->watom->what & (TYPETW | TYPEPW))) \
